Fix display of breaks in the backend for working plans stored in database with Monday as the first weekday (EA! Issue #506).
Prior commit 00372f2f1af96df0d4545053ae6f27834e109ea6, the first weekday was Monday.
After this commit, it is set to Sunday and the display of breaks in the backend calendar is broken for former working plans.

The fix consists in reordering the working plan elements with Sunday as the first day prior displaying the backend calendar.
In addition, the working plan elements are also reordered when displaying 'Settings/Business Logic' and 'Users/Providers' tabs.
This is to ensure that breaks are always displayed with Sunday first, like done for the Working Plan in those tabs, for consistency purpose..