WordPress Functions Hook-in Modification
This allows full hook in to the wordpress functions library to better integrate with the existing Wordpress functions library. It allows WordPress user database to cooperate with Easy!Appointments.  For example if you have Easy!Appointments sitting behind a Wordpress Login, the following will auto fill: user name, email, and phone number (if you have added phone to the Wordpress registration routine).

With this hook in the Wordpress theme text characteristics are then reflected in Easy!Appointments.

Also with this hook in, Wordpress functions can be used within Easy!Appointments.  This allows me to use a wordpress PayPal and Invoicing plug in with Easy!Appointments. I will demonstrate that later.