Date Picker Notice and Formatting Changes
These changes center up the time slots when in mobile designe mode.

In the standard build the notice that there are no availble appointments only pops up if the calendar loads on an unavailable date.  I have also added some code to notify customers when clicking any date that is unavailable that there are no appointments for that date.