Adding a reminders utility
This reminders utility works with a chron job or similar to launch the script found in the application/controllers/cli/Reminders.php file.  The frequency of reminders is set in the settings page on the back end.  By default the reminders go out 3 days and 1 day before the appointment.  I have included a simple HowToUseCLI file to explain how to use the CLI feature of Codeigniter.