Added 2 options for Availabilities Types and Time interval
In user settings I am adding two more options for Availabilities Types: On the quarter and on the half.  These force the appointments to sit on the quarter hours or on the half hours.  They make it possible to have padding  between appointments.  For example if on the quarter is selected and you have a 15 min appointment (or 30, 45, or 60 min and so on) there will be a 15 min pad of free time to the next appointment.  If you have a 20 min appointment you willl have a 10 min pad of free time etc.. It is best if you are going to use "On the quarter" or "On the half" to stick to one or the other if you use either one of them at all.  It is not a perfect padding option but it is the easiest for me for now and it works for my business.

I have also added a varable on the settings page to adjust the "Time Interval For Fixed Appointments".  The default is 15 min.   This makes available appointment spots appear in 15 min incraments.  Now you can adjust it to anything you like.  I prefer 60 min increments.