Waiting list function added
I have added a setting to turn on or off the waiting list from the view. I have added a place in settings where you can modify the message for the waiting list.  Clients can register for the waiting list on the front end if activated.  I have added a script in /application/controllers/cli/Waitinglist.php that can be launched by Cron Job or equivalant on the server.  Instructions for using the Waitinglist.php script are found in this text file: /volume1/web/wordpress/easytest/application/controllers/cli/HowToUseCLI.txt