Announcement script
This is a rugh and dirty script I use to make announcements to clients that I will be taking time off.  Before I delete them from the system I update this script with the days I will be taking off and other details.  It then will send an email to all my clients who were scheduled for those days informing them of when I will be gone.  I have not added any script to delete the appointments programmically.  I should.  But for now I am doing it by hand.