Preview Page Added
This is a preview page that I use on my front end so those not logged in can see my availability, but they cannot book on this page.  In order to book I require clients to log in.  I also made some other corrections along the way to the language file for other things.