Add Special Catagory for Specific Clients Feature
I have added a feature where I can make some catagories of services only available to specific clients, or I can make alternate charge schedule for specific clients.  So if some clients have a $20 copay, they can book a session for that price on the front end if I set up that category for the client(s). Otherwise they only have the option of paying my full fee.

INSTRUCTIONS:
Special Category option is only available within WordPress behind a log in for the client.  Within the services>catagories tab create a new category as usual.  When creating a special category it is helpful to add a prefix to the name like "Spc-".  There is a new field called "Special Category".  Enter a code for the special category here (up to 32 charactors).  Assign services to that category that you want available only to specific clients in the services tab.  In the customer's tab you will find a "Special Category" field there.  Enter the code for the special category there.  Now, when the client logs into WordPress to book an appointment they will see the special category in addition to the standard categories.