Update Appointments.php
I think that this controller has a little bug. When you have multiple attendants in a service and you are trying to modify or cancel an appointment using the link that is given in the confirmation email, the system shows an error.
I've been debugging the project and finally I have found that the format of the date can be wrong depending on your configuration. With the modification at line 398 the problem is solved.