Retain linebreaks in the service description
Our org uses EasyAppointments (I'm not an admin, just a org-side user) and would like to add extensive descriptions with enumerations etc. in the service description, however these don't seem to be retained from the input text to the display text (that a user sees). 

From my limited knowledge regarding php/html and a bit of research, I think adding nl2br to line 178 would do the trick, but I'm not sure.
Greatly appreciate input on this!